pyproject.toml
[project]
name = "realtime-api-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "websockets>=14.1",
    "chainlit>=2.0.3",
    "aiohttp>=3.11.11",
    "aiohttp-cors>=0.7.0",
    "azure-identity>=1.19.0",
    "python-dotenv>=1.0.1",
    "numpy>=2.2.2",
]
